Move functions from _util.ts to validation.ts
